requirements.txt
pandas~=1.3.4
numpy~=1.21.4
scikit-learn~=1.0.1
mysql-connector-python~=8.0.27
flask
flask-restful
seaborn